requirements.txt
spotipy=2.24.0
requests=2.32.3
python-dotenv=1.0.1
python=3.9.13
db-sqlite3=1.2.1
